Refactor create position
