+++ src/components/PositionModal/BorrowForms/CreateForm.tsx
@@ -9,20 +9,25 @@ import { commifyToDecimalPlaces, floor2, floor4, minCollRatioWithBuffer } from '
 import { Alert } from '../../Alert';
 import { NumericInput } from '../../NumericInput/NumericInput';
 import { Slider } from '../../Slider/Slider';
+import { buildBuyCollateralAndModifyDebtArgs, sendTransaction } from '../../../actions';
+import { useUserData } from '../../../state/queries/useUserData';
+import { chain as chains, useAccount, useNetwork } from 'wagmi';
+import { useAddRecentTransaction } from '@rainbow-me/rainbowkit';
+import useSoftReset from '../../../hooks/useSoftReset';
 
 const CreateForm = ({
   onClose,
-  // TODO: refactor out into react query mutations / store actions
-  createPosition,
   setUnderlierAllowanceForProxy,
   unsetUnderlierAllowanceForProxy,
 }: {
   onClose: () => void,
-  // TODO: refactor out into react query mutations / store actions
-  createPosition: (deltaCollateral: BigNumber, deltaDebt: BigNumber, underlier: BigNumber) => any;
   setUnderlierAllowanceForProxy: (fiat: any, amount: BigNumber) => any,
   unsetUnderlierAllowanceForProxy: (fiat: any) => any,
 }) => {
+  const { chain } = useNetwork();
+  const { address } = useAccount();
+  const addRecentTransaction = useAddRecentTransaction();
+  const softReset = useSoftReset();
   const modifyPositionData = useStore((state) => state.modifyPositionData);
   const {
     collateralType: {
@@ -38,7 +43,11 @@ const CreateForm = ({
   const hasProxy = useStore(state => state.hasProxy);
   const disableActions = useStore((state) => state.disableActions);
   const transactionData = useStore((state => state.transactionData));
+  const user = useStore((state) => state.user);
   const { action: currentTxAction } = transactionData;
+
+  const { data: userData } = useUserData(fiat, chain?.id ?? chains.mainnet.id, address ?? '');
+  const { proxies } = userData as any;
 
   const borrowStore = useBorrowStore(
     React.useCallback(
@@ -59,7 +68,18 @@ const CreateForm = ({
 
   const [submitError, setSubmitError] = React.useState('');
 
-  const minCollRatio = useMemo(() => minCollRatioWithBuffer(liquidationRatio), [liquidationRatio])
+  const minCollRatio = useMemo(() => minCollRatioWithBuffer(liquidationRatio), [liquidationRatio]);
+
+  const createPosition = async (deltaCollateral: BigNumber, deltaDebt: BigNumber, underlier: BigNumber) => {
+    const args = buildBuyCollateralAndModifyDebtArgs(
+      fiat, user, proxies, modifyPositionData.collateralType, deltaCollateral, deltaDebt, underlier
+    );
+    const response = await sendTransaction(
+      fiat, true, proxies[0], 'createPosition', args.contract, args.methodName, ...args.methodArgs
+    );
+    addRecentTransaction({ hash: response.transactionHash, description: 'Create position' });
+    softReset();
+  }
 
   if (
     !modifyPositionData.collateralType ||
