+++ src/components/PositionModal/PositionModal.tsx
@@ -20,7 +20,6 @@ const enum Mode {
 }
 
 interface PositionModalProps {
-  createPosition: (deltaCollateral: BigNumber, deltaDebt: BigNumber, underlier: BigNumber) => any;
   buyCollateralAndModifyDebt: (deltaCollateral: BigNumber, deltaDebt: BigNumber, underlier: BigNumber) => any;
   sellCollateralAndModifyDebt: (deltaCollateral: BigNumber, deltaDebt: BigNumber, underlier: BigNumber) => any;
   redeemCollateralAndModifyDebt: (deltaCollateral: BigNumber, deltaDebt: BigNumber) => any;
@@ -161,7 +160,6 @@ const PositionModalBody = (props: PositionModalProps) => {
       return !!selectedCollateralTypeId && actionMode === Mode.CREATE
           ? <CreateForm
               onClose={props.onClose}
-              createPosition={props.createPosition}
               setUnderlierAllowanceForProxy={props.setUnderlierAllowanceForProxy}
               unsetUnderlierAllowanceForProxy={props.unsetUnderlierAllowanceForProxy}
           />
